+++ src/main/java/ru/mail/polis/dao/s3ponia/Table.java
@@ -58,7 +58,6 @@ public interface Table extends Closeable {
         private final long deadFlagTimeStamp;
         private final int generation;
 
-
         /**
          * Value constructor.
          *
