+++ src/main/java/ru/mail/polis/s3ponia/Utility.java
@@ -1,8 +1,6 @@
 package ru.mail.polis.s3ponia;
 
 import com.google.common.base.Splitter;
-import com.google.common.collect.Comparators;
-import jdk.jshell.execution.Util;
 import one.nio.http.HttpClient;
 import one.nio.http.HttpServerConfig;
 import one.nio.http.Request;
@@ -12,7 +10,6 @@ import one.nio.server.AcceptorConfig;
 import org.jetbrains.annotations.NotNull;
 import ru.mail.polis.dao.s3ponia.Table;
 
-import java.net.http.HttpResponse;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -21,7 +18,6 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
